Fully functional BTree implementation.
This BTree is optimized as far as I could manage, passes all given unit
tests, and has all documentation.

We will still want to write more unit tests around the methods that are
not directly tested, which are those that are not apart of the
BTreeInterface, excepting `getSortedKeyArray`, which was wielded despite
that fact anyway.

Note that the the error handling in the constructors is a bit much, and
they feel needlessly complex. This is due to two factors:

1. The tests demanded that these constructors throw `BTreeExceptions`
   instead of what they would otherwise throw: `IOExceptions`. This
means theres a lot of try/catch statements.
2. The constructor that takes a `fileName` breaks clean code principles
   by trying to do two things: read an existing file, or barring that,
create a new one. We already have a constructor that creates a new one,
but the simple act of having to check first whether the file exists
means that Java won't let us reuse the other constructor, as `this()`
is required to be called first thing. So we have to duplicate the other
constructor's code almost line for line to initialize properly.

The changes I made to Derek's algorithims were the following:

- The loops Derek wrote stopped short of processing the final value in
most instances. He was copying the pseudocode given in the notes, and
the syntax was the "from-to inclusive" type. Easy to see, easy to fix.
- Because we're expecting duplicates to be entered and we're to count
  frequencies, there were times where we had to check in the algorithms
if a duplicate was being inserted, and stop short if it was. This has
been inserted.
- There was an instance in insertion where splitting the child in a
  different scope wouldn't update the nodes we had in the calling scope,
and that forces us to do an additional disk read to get the updated
nodes after the child has been modified.

Other notable changes were:

- I began tracking height in the metadata. I wasn't convinced there was
  a good way to calculate height based on node count. But then I learned
from the textbook that the height of a BTree only increases when the
root is split, which made for easy tracking. It now reflects that.
- Subsequence length to decode the `long` keys is nowhere exposed to
the BTree and its objects. To adjust for this until we get further
review from Amit, I've given `TreeObject` a class variable that must be
set before using `dumpToDisk`, which needs to know subseuqence length
before it can interpret its keys as ASCII characters.

Overall, this should be conceptually sound and functional. But I stress
again we'll want to wrap it in more unit tests than currently exist.