Tested and bugfixed `BTreeInOrderIterator`.
The title of this commit says it all. It was previously communicated
that the `BTreeInOrderIterator` had a potential for bugs. And knowing
that, it was a little unnerving that the project-given unit tests still
passed. So this commit has added a test file that broadly tests it for
accuracy with differing minimum degree trees, counts, and loads. During
its creation, the logic of the `next` method changed, and a couple of
helper methods were also created.

Note that this `BTreeInOrderIterator` does not handle the case of
concurrent modification. That is perhaps worth adding, but given the
project's intentions is to completely build up the tree before doing any
kind of reading or checking, we may be just fine.