Added `GeneBankSubsequenceIterator.java` + tests.
This new class does what you'd expect it to: returns DNA subsequences as
strings given the project specification. A rough outline is as follows:

- When constructed, it will seek to the line after the first "ORIGIN"
  declaration in the file and advance its position until the next
character is a valid DNA base. If one is not found, it will error.
- It was easiest to design these algorithms if, upon construction and
  each call to next, we went ahead and checked if there was another one
for the sake of `hasNext`. In doing so, we get everything in a state to
simply read the next valid character when `next` is invoked again.
- Any call to next will return the next subsequence in the file,
  resetting the subsequence when "n" characters are encountered or when
the ORIGIN block ends. Line breaks do NOT cause a reset in the
subsequence. This seemed to be the intention of the project
specification.

This class only implements `next` and `hasNext`, making it a bare-bone
Iterator we can use with our programs.

Note that this Iterator will return Strings of the subsequence length
you specify upon construction.