Added `BTreeSQLiteDBBuilder.java`.
This class exposes one static method, `create`, that takes a reference
to a `BTree` and a `Path` that corresponds to the `.gbk` file that
created it. It creates and populates a SQLite database that contains all
the subsequence encodings and their frequency counts, per the project
specification.

Since the main application for the database was going to be searching, I
elected to insert the subsequences into the database as a 64 byte `long`
and I put an index on it. Those two factors should make searches
extremely fast and efficient.

This class hasn't been tested, as we need the program that creates a
BTree from a given `.gbk` file ready in order to wield it; at that
point, we will test.