BTree tested more; All unit tests passing.
In the work for this commit, one bug was identified when inserting a
duplicate into a leaf node: Before checking for duplicates, space was
made in the node by copying all references to `TreeObjects` forward in
the node by one. However, if a duplicate was found, this copy wasn't
undone, and leaving a lingering extra reference in the node. This was
discovered while testing the `dumpToFile` method, which doubled the
print of the `TreeObject` that now had two references sitting in a given
node.

Also added code to parse the arguments for the create program in order
to get the last failing unit test passing.

Something worth noting on that front is that the project usage given in
the README suggests that the program uses long arguments for what are
truly positional/required arguments. This breaks with the POSIX
standard. The given tests, however, expects to be able to successfully
test positional arguments without options, and this does align with this
POSIX standard. So I wrote logic to try parsing as positional arguments
first, and, upon that failure, I default to POSIX style parsing. The
latter is not tested by the given unit tests.