Implemented the cache.
One of the complicating aspects of the copy-and-paste job we're supposed
to do from our old project is that the `Cache` uses the objects
themselves as keys, which is generally not a great design choice because
it makes you ask for objects for which you already have references. For
example, with the `Webpages` we originally did, we stored them in the
cache keyed by the URL of the `Webpage`. But where was the ultimately
stored? In the `WebPage` itself. I shuddered and then stopped thinking
about it because ultimately I just had to pass tests that were given.
But it created a more pernicious problem here.

We had designed the `BTreeNode` to be lightweight; it only hosted its
key values and its child positions, something akin to a struct. It also
had some static utility methods to calculate size, but that wasn't
really derived from the contents of the node. Ultimately, we were left
without content in the node that could self-identify it compared to all
other nodes. (Well, not exactly. I briefly considered that we could
concatenate a string of all the values, since duplicates aren't present
in the `BTree`, just its frequency counts. But computationally, that's
very expensive when everything is a long, and you could have >100 of
them in a given node.)

I opted instead to have the BTree manage the manual setting of the key
during creation and reconsitution of nodes. You'll see that the changes
to `BTree.java` include the following:

- A new field `cache` whose default is `null`. It does not get set
  unless you use the following constructor...
- A new constructor that assumes you want a cache and asks for its
  capacity. This sets the `cache` field and then duplicates the
constructor code from the similar constructor that has already been
verified to be working.
- Logic within `diskRead`, and anywhere where `new BTreeNode(...)`
  or `BTreeNode.fromByteBuffer(...)` appears, which are acts of
reconstitution.

The nature of the refactor was surgicial, with the ultimate goal to do
the least disruption to the codebase.

Upon any call to `BTree.diskRead(long position)`, the given position
will be checked in the cache, as that's the key. If it is found, it
returns early. If not, it reconstitutes from disk, sets the key on the
object as this scope is aware of its position, and then returns.