Finalized CreateBTree program.
This really took a lot of the groundwork that Derek laid and pushed it
over the finish line.

The biggest change was to the create program itself. After some
clarification about the CLI on Piazza, the parsing logic became much
more succinct and straightforward, and I was able to eliminate the "dual
path" approach we were taking previously: where we'd try to parse the
arguments as positional first, and then named. Now they're only ever
going to be named. The logic was centralized in
`GeneBankCreateBTreeArguments` within a factory method that can take in
a String[] type (the command line arguments themselves as parsed by the
JVM). This will return an instance of the class ready for use. As there
was one method that the tests expected to interface with on the
`GeneBankCreateBTree.java` program to make sure arguments parsed fine,
it routes rightly to `GeneBankCreateBTreeArguments.java`. Tests have
also been added to assure everything parses correctly.

Other notable changes include:

- Profiled and fixed an IO issue in the `BTree` that was consuming
  an unacceptable amount of runtime by trying to keep the fileChannel
synced with disk. Now it only syncs upon call to `BTree.finishUp`.
- Fixed an issue with `BTree.dumpToFile` where it wasn't flushing the
  `OutputStreamWriter`. Now dump files populate correctly.
- Fixed issues with the `BTree` file channel closing prematurely.
- Removed an unnecessary filename check in `BTreeSQLiteDBBuilder.java`.

I ran all the different creation scenarios for files test0.gbk and
test5.gbk, ran the script to compare the dump files, and **everything
matched exactly!** So in terms of testing, we're looking pristine.

Also notable is that, without a cache, our software is more performant
than the professors runtime by a factor of almost 2x! We'll need to
profile a bit more, but I teased out all the blantanly obvious speed
inefficiencies with this commit.

TODO: We still have to create a quick branching of BTree construction
once we get the cache implemented. That can be done after PR #51 is
merged into master.