Tests now run create and dumpfile-check scripts.
I added a Gradle task that will run the "create-btress.sh" and
"check-dumpfiles.sh" scripts on the test0.gbk file when tests are
invoked. These can be though about as an integration test given by the
project.

Creating this task is useful because it will be invoked by the Github
Actions pipeline in future PRs. If we make an update and this test
starts failing, we'll know we made a change that gets us farther away
from a good grade rather than closer.