Finished cache implementation.
After some bad performance with the Project 1 implementation of a cache
and some back and forth with Amit on Piazza, I learned that the
Professors' solution utilized a `LinkedHashMap` as their LRU cache
instead of using a similar solution from Project 1. Amit gave me
permission to do the same here.

A high-level overview of the changes in this commit are as follows:

- Decoupling the node position in the file from the fileChannel, so
  nodes could be in the cache without being in disk and yet all nodes
would still get distinct positions with proper sizes.
- Removing the old cache implementation and code to make way for the use
  of the LinkedHashMap.
- Refactoring out the actual disk write operation out of `diskWrite`
  because there are a handful of situations where the root needs to be
written to disk, such as in a constructor or a root split.
- Making the `diskWrite` and `diskRead` methods utilize the cache if its
  available.