Search program now using cache.
This required adding a new constructor to load a file from disk. In that
sense, the two constructors without a cache now mirror two constructors
with a cache.

I refactored out the creation of the `LinkedHashMap` construction + the
overridden method to not duplicate code in two places.

All test have ran and passed. All integration scripts have been ran on
`test0.gbk` and `test5.gbk` and have passed.