Wrote more comprehensive logging.
Also added some detail to `runtimes.md` and increased the sleep time for
`run-all.sh` between test results.

The create program now presents a bunch of summary numbers about the
BTree immediately after creation, and logs when it's dumping as well as
when it's finalizing.

The search program now presents search results in a tabular format that
is meant to convey the information and its interrelation as quickly as
possible.