Filled out the majority of README-submission.md.
All that's left is Derek's reflection, which can be committed to this
branch before merging to master. The reflection also doesn't have be
nearly as long as mine; Derek has gotten to know me over the course of
the project, and as I'm sure he's intuited, I certainly always have a
lot of thoughts and things to say, haha.